Improved GUI interaction, created nice_plots.py
GUI:
Aligned units in the time_array, it is now fixed to seconds even when plotting. This fixed a bug where clicking in the plots of the GUI caused the wrong spectrum to be shown, because the plot units were hours but the time_array was seconds.

nice_plots.py:
Created this file with the idea to create nice plots using matplotlib. Still figuring out how to make matplotlib use Latex, didn't work on Mac OSX.