GUI.py: Included power spectrum, nice_plots.py: Latex included
GUI.py:
Removed the plot of the linewidth histogram and replaced it with the power spectrum of the frequency over time data (resulting from the fits).
Also tried to only save the relevant Lorentz parameters (pos, lw). Does not work however because they are all needed for the function where the successful fit plotted when clicking in the frequency or linewidth plot.

nice_plots.py:
Completed plot template, including Latex font in ticks and axes labels.